mac_chatbot/requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
google-cloud-storage>=2.14.0
